% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{ACFdetrended}
\alias{ACFdetrended}
\title{Modified autocorrelation function on detrended data}
\usage{
ACFdetrended(x, y, sig.level = 0.95)
}
\arguments{
\item{x}{numeric vector of timepoints}

\item{y}{numeric vector of observations}

\item{sig.level}{level of significance to test at.}
}
\value{
a vector of 2 values describing lag-1 autocorrelation of the
detrended series, and whether the autocorrelation is significant.
}
\description{
Detrends a time-series using Theil-Sen and computes the lag-1 autocorrelation,
and returns its value and whether it is significant to a given level.
}
