% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{ACFmod}
\alias{ACFmod}
\title{Modified autocorrelation function}
\usage{
ACFmod(x, sig.level = 0.95)
}
\arguments{
\item{x}{a vector of observations}

\item{sig.level}{value in [0,1] describing level of significance}
}
\value{
a vector containing the lag-1 autocorrelation, and a boolean indicating
 significance.
}
\description{
Computes the lag-1 autocorrelation of a time-series, and returns its value and
whether it is significant to a given level.
}
