% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{amaxPlot}
\alias{amaxPlot}
\title{Plotting of AMAX series with trend lines}
\usage{
amaxPlot(x, param, start, end, rp = 2, NST = FALSE)
}
\arguments{
\item{x}{time-series dataframe of year and flow}

\item{param}{vector of GLO parameters (loc, sca, sha) for stationary distributions and (loc_int, loc_slope, sca, sha)
for non-stationary distributions}

\item{start}{start year of plot}

\item{end}{end year of plot}

\item{rp}{vector of return periods (in years for AMAX data)}

\item{NST}{if true, plots non-stationary return period lines.}
}
\value{
a plot with the time-series and lines indicating magnitude of
events with given return periods over time.
}
\description{
Uses a time-series and fitted parameters to add lines indicating estimated
magnitudes of events of given return periods.
}
