% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoAxis}
\alias{autoAxis}
\title{Add automatically scaled horizontal date axis.}
\usage{
autoAxis(series, ...)
}
\arguments{
\item{series}{vector of datetime objects (readable by \code{as.Date})}

\item{...}{additional \code{par} arguments to be passed to \code{axis}.
Can include any argument passed to axis except \code{side}, \code{at}, \code{labels} and \code{tick}.}
}
\value{
Adds axis to existing plot in current device.
}
\description{
Using a provided time series, this functions adds a horizontal datetime axis
to an existing plot in the same fashion as the \code{axis} function. Ticks and
labels are automatically determined. Additional graphical \code{par} arguments
can be passed to customise apperance.
}
\examples{
\dontrun{
    ts <- seq(as.Date("01-01-1900"),as.Date("01-01-1902"),by="day")
    autoAxis(series = ts, cex.axis=0.6, las=1)
}

}
