% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{detrend}
\alias{detrend}
\title{Detrended time-series using Theil Sen estimates of trend.}
\usage{
detrend(x, y)
}
\arguments{
\item{x}{numeric vector of timepoints}

\item{y}{numeric vector of observations}
}
\value{
a list containing: three dataframes,
   \itemize{
   \item a dataframe with two columns, timepoints and detrended values under linear regression
   \item a dataframe with two columns, timepoints and residuals under linear regression
   \item a dataframe with two columns, timepoints and detrended values under Theil-Sen estimation.
   }
}
\description{
Given a time-series with values and time points, returns detrended time-series and
residuals from linear regression. Also returns the a dataframe of the detrended time-series
under Theil-Sen.
}
