% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potextract.R
\name{extractPOT}
\alias{extractPOT}
\title{Extract Peak-over-Threshold series from timeseries (NRFA criteria)}
\usage{
extractPOT(series, datetime = NULL, threshold = 0, timeOfRise = 0)
}
\arguments{
\item{series}{Either a vector of flow magnitudes, or a dataframe with two
columns, magnitude and datetime.}

\item{datetime}{If series is not a dataframe, this is a vector of datetimes
corresponding to the events in \code{series} vector. If not supplied,
and no datetime supplied, ordinal values are assumed (1,2,...,N).}

\item{threshold}{positive number describing minimum size of peak event. If
not supplied, takes the 90th percentile of the data.}

\item{timeOfRise}{positive number describing time of rise at the location of
the measurements.}
}
\value{
a list containing:
    \itemize{
    \item a dataframe of the peak events, including datetime and magnitude
    \item a boolean vector the length of the original series indicating which
        measurements are peaks.
    }
}
\description{
Using a given time series, a POT series is returned using the NRFA criteria:
\itemize{
\item Peaks must be seperated by a time interval at least 3 times as long as the
estimated time to rise at the location of interest.
\item Between peaks, the trough must be no more than two-thirds the size of
either peak.
\item All peaks must exceed a given threshold: current standard is to choose a
threshold which is expected to permit, on average, five events per year.
}
This is based on the work of Ilaria Prosdocimi and the packages \code{pastecs}
and \code{ilaProsUtils}.
}
\examples{
\dontrun{
    xt <- (arima.sim(list(order=c(3,0,0), ar=c(0.5,0.25,0.125)), n=100) + 2)^2
    dt <- as.Date(1:100, origin="2000-01-01")
    extractPOT(xt, dt, threshold = 20, timeOfRise = 3)
}

}
