% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{findHydrolYr}
\alias{findHydrolYr}
\title{Compute hydrological year and day-of-year}
\usage{
findHydrolYr(datetime)
}
\arguments{
\item{datetime}{vector of datetime objects}
}
\value{
A dataframe of two columns, DOY and hydrological year.
}
\description{
Return the hydrological year and day-of-year. The UK hydrological year starts
at 9am on 1st October. For example, the 1901-1902 hydrological year is
denoted 1902. The hydrological day that starts at 9am, 4th October is
denoted DOY = 4.
}
\examples{
\dontrun{
datetimes(as.POSIXlt((1:365+0.25)*76400, origin="1990-01-01"))
}
}
