% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idConverter.R
\name{idConverter}
\alias{idConverter}
\title{Convert between different station identifiers.}
\usage{
idConverter(num, source = c("NRFA", "EA", "SEPA"))
}
\arguments{
\item{num}{vector of numbers or strings containing identifier(s) of station
(all should come from same source)}

\item{source}{string indicating source of provided identifier(s), not target.}
}
\value{
the identifier(s) from the other data source not selected in 'source'.
    If not found, returns NA.
}
\description{
Given a station identifier, and its source (either NRFA or EA/SEPA), this function
provides the identifier designated by the other organisation. idConverter
assumes that any given station is uniquely monitored by EA or SEPA, not both.
}
\examples{
\dontrun{
idConverter(39071, "NRFA")  # NRFA to EA/SEPA
idConverter("0130TH", "EA")  # EA to NRFA
}

}
