% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idConverter.R
\name{idToRef}
\alias{idToRef}
\title{Obtain reference for river flow API page.}
\usage{
idToRef(id)
}
\arguments{
\item{id}{EA station number.}
}
\value{
string to find flow from API
}
\description{
For EA stations for which daily flow is provided, this provides the reference
string for the API, given the EA identifier.
}
\examples{
\dontrun{
idToRef("SS50F007")
}
}
