% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ts.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata from river flow API.}
\usage{
import_metadata(ids, dat, org = c("NRFA", "EA", "SEPA", "COSMOS"))
}
\arguments{
\item{ids}{identifier for stations (not EA refs)}

\item{dat}{string indicating datatype, as written in metadata.}

\item{org}{organisation from whom the data is obtained.}
}
\value{
a list containing:
\itemize{
\item id - measuring authority station identifier
\item ref - API reference string
\item name - station name
\item organisation
\item station aliases under different organisations
\item datatype - list of descriptors of data
\item startDate - character string of first record
\item dataUrl - string of URL to obtain data from API directly.
}
     If not found, returns NA for each such station.
}
\description{
Using the river flow/rainfall API, station information can be extracted for
single or multiple sites. All data must be of the same type from the same
organisation.
}
\examples{
\dontrun{
import_metadata(ids=c("SX67F051", "SS50F007"), org="EA", dat="gdf")
import_metadata(ids="SX67F051", org="EA", dat="gdf")
}

}
