% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ts.R
\name{importTimeSeries}
\alias{importTimeSeries}
\title{Import time series from river flow API.}
\usage{
importTimeSeries(
  ids,
  dat,
  org = c("NRFA", "EA", "SEPA", "COSMOS"),
  startDate = NULL,
  endDate = NULL,
  metadata = FALSE,
  datetime = TRUE
)
}
\arguments{
\item{ids}{identifier for stations (not EA refs)}

\item{dat}{string indicating datatype, as written in metadata.}

\item{org}{organisation from whom the data is obtained.}

\item{startDate}{string of the form \code{YYYY-MM-DD} to indicate start of
period desired, or single date. Whole record given if no startDate
provided.}

\item{endDate}{string of the form \code{YYYY-MM-DD} to indicate end of
period desired. If no startDate provided, this is ignored.}

\item{metadata}{if \code{TRUE}, returns metadata for each station selected.}

\item{datetime}{if \code{TRUE}, converts text datetime column into POSIXlt.}
}
\value{
a dataframe containing the dates and magnitudes of the selected
    data. If multiple stations selected, the dataframes are contained in a
    named list. If metadata is true, each station will consist of a list
    containing \code{detail} and \code{data}.
    If not found, returns NA for each such station.
}
\description{
Using the river flow/rainfall API, time series can be extracted, either
selecting single dates, periods of record, or entire records for single or
multiple sites. Metadata can also be returned for stations in the dataset.
All data must be of the same type from the same organisation over the
same period.
}
\examples{
\dontrun{
importTimeSeries(ids=c("SX67F051", "SS50F007"), org="EA", dat="gdf",
    startDate="2017-01-01", endDate="2017-02-01")
importTimeSeries(ids="SX67F051", org="EA", dat="gdf", metadata=T)
}

}
