% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{importTriangle}
\alias{importTriangle}
\title{import data from API for triangle plotting}
\usage{
importTriangle(st_no, start, end)
}
\arguments{
\item{st_no}{string or numerical station NRFA identifier}

\item{start}{start year}

\item{end}{end year}
}
\description{
This takes AMAX data from the NRFA API, and reformats it into shape for a
multi-temporal analysis of trend. The analysis needs at least 26 years of data as
it uses a 25-year rolling window.
}
