% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{KendallZScore}
\alias{KendallZScore}
\title{Return Mann-Kendall test statistics}
\usage{
KendallZScore(x)
}
\arguments{
\item{x}{vector of observations.}
}
\value{
Z, the test statistic for Mann-Kendall
}
\description{
Given a vector of observations, return the test statistic for the adjusted
Mann-Kendall test as in the \code{Kendall} package. This allows for repeated
values.
}
