% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{parglo_mle_tse}
\alias{parglo_mle_tse}
\title{Maximum likelihood estimation of non-stationary distributions using TSE}
\usage{
parglo_mle_tse(amax, years, init_params)
}
\arguments{
\item{amax}{vector of AMAX values}

\item{years}{vector of timepoints corresponding to \code{amax}}

\item{init_params}{vector of initial parameter estimates \eqn{(loc_{int}, loc_{slope}, sca, sha)}}
}
\value{
vector of parameters (loc, TSE, sha, sca)
}
\description{
fits the GLO parameters as described in FEH Vol III,
assuming a linear trend in the location parameter only, which is
fixed to be the Theil-Sen Estimator.
}
