% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rainFlowPlot}
\alias{rainFlowPlot}
\title{Plot rainfall and flow time series on the same plot.}
\usage{
rainFlowPlot(rainID = NULL, flowID = NULL, rainOrg = NULL,
  flowOrg = NULL, rainDat = "cdr", flowDat = "gdf",
  startDate = NULL, endDate = NULL, filepath = NULL)
}
\arguments{
\item{rainID}{identifier for station recording rainfall}

\item{flowID}{identifier for station recording flow}

\item{rainOrg}{organisation responsible for the station in \code{rainID}}

\item{flowOrg}{organisation responsible for the station in \code{flowID}}

\item{rainDat}{rainfall measurement designator as described in metadata}

\item{flowDat}{flow measurement designator as described in metadata}

\item{startDate}{string of the form \code{YYYY-MM-DD} to indicate start of
period desired, or single date. Whole record given if no startDate
provided.}

\item{endDate}{string of the form \code{YYYY-MM-DD} to indicate end of
period desired. If no startDate provided, this is ignored.}

\item{filepath}{if provided, this gives the path to save a .png image.
".'png" must be included at the end of the filepath.}
}
\value{
plots figure to open device unless \code{filepath} provided, in which
    case a .png file is saved instead.
}
\description{
Given an identifier for a rain measurement and/or a flow gauging station,
this function will plot the combined plot of rainfall and river flow, or a
single plot of the desired time series.
}
\examples{
\dontrun{
rainFlowPlot(rainID = "234176", flowID = "9005",
             rainOrg = "SEPA", flowOrg = "NRFA",
             startDate = "2017-07-01", endDate = "2018-07-01")
}

}
