% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{relativeTheilSen}
\alias{relativeTheilSen}
\title{Relative Theil Sen Estimator}
\usage{
relativeTheilSen(x, y)
}
\arguments{
\item{x}{numeric vector of timepoints}

\item{y}{numeric vector of observations}
}
\value{
the value of relative TSE.
}
\description{
Computes the relative Theil-Sen Estimator of a time-series which is given by \eqn{\xi N / \mu}
where \eqn{\xi} is the standard TSE, and \eqn{\mu} the mean of the time-series.
}
