% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idConverter.R
\name{stationList}
\alias{stationList}
\title{Gives a list of stations from the relevant organisation.}
\usage{
stationList(org = c("EA", "NRFA", "SEPA", "COSMOS"))
}
\arguments{
\item{org}{organisation to obtain list of stations from.}
}
\value{
dataframe containing station names, ids and other important information.
}
\description{
Returns a dataframe of stations from a selected organisation.
Columns include NRFA station number, latitude, longitude, id, JSON
reference, and corresponding identifiers under other organisations.
}
\examples{
\dontrun{
stationList("NRFA")
}

}
