% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{trianglePlot}
\alias{trianglePlot}
\title{Triangle plotting of trends for all observed time windows}
\usage{
trianglePlot(MKZ, ts)
}
\arguments{
\item{MKZ}{matrix of Mann-Kendall Z-scores}

\item{ts}{time-series dataframe of year and flow}
}
\value{
triangle plot with start year along the side, end year along the bottom.

Each gridsquare shows the Mann-Kendall statistic for that time period.
Points which are statistically significant at 95% are highlighted.
}
\description{
This function uses the output from \code{triangleTrends} and plots the triangle
plot.
}
