% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfunctions.R
\name{triangleTrend}
\alias{triangleTrend}
\title{Triangle plotting of trends for all observed time windows}
\usage{
triangleTrend(ts, startyr, endyr)
}
\arguments{
\item{ts}{a two column data.frame object of year and flow}

\item{startyr}{start year for triangle matrices}

\item{endyr}{end year for triangle matrices}
}
\value{
list of 2 matrices, MKZ values and TSE values.
}
\description{
This function takes the data.frame of amax flow, and returns a matrices of
Mann-Kendall test statistics and Theil-Sen estimates of trend for every choice
of start year and end year (with a minimum of 26 years of data)
}
\section{Information about flags}{

\enumerate{
\item for the years outside the range of data for each station give the value -100 \\
\item when there are less than 27 total years (NA included) give the value -200 \\
\item when there are: \itemize{
   \item more than 10% missing data or less than 27 actual data (NA excluded),\\
   \item or more than 2 missing values in the beginning of the subset,\\
   \item or more than 1 missing value in the end of the subset, then give the value -300 }\\
\item when end year is smaller than start year give the value -400
}
}

