% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ts.R
\name{ts_reformat}
\alias{ts_reformat}
\title{Reformats a time series to have datetime objects.}
\usage{
ts_reformat(ts)
}
\arguments{
\item{ts}{time series data.frame object of two columns: datetime
(string of datetimes in form \code{YYYY-MM-DDTHH:MM:SSZ}) and
data (numeric).}
}
\value{
data.frame with replaced datetime column containing equivalent
    POSIXlt objects.
}
\description{
Converts a data.frame with strings for datetimes into one with POSIXlt date
objects.
}
