% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idConverter.R
\name{idConverter}
\alias{idConverter}
\title{Convert between different station identifiers.}
\usage{
idConverter(id, org = c("NRFA", "EA", "SEPA"))
}
\arguments{
\item{id}{vector of numbers or strings containing identifier(s) of station
(all should come from same org)}

\item{org}{string indicating source of provided identifier(s), not target.}
}
\value{
the identifier(s) from the other data providers not selected in 'org'.
    If not found, returns NA.
}
\description{
Given a station identifier, and its source (either NRFA or EA/SEPA), this function
provides the identifier designated by the other organisation. idConverter
assumes that any given station is uniquely monitored by EA or SEPA, not both.
}
\examples{
\dontrun{
idConverter(39072, "NRFA")  # NRFA to EA/SEPA
idConverter("2700TH", "EA")  # EA to NRFA
}

}
