% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idConverter.R
\name{idToRef}
\alias{idToRef}
\title{Obtain station reference for river flow API.}
\usage{
idToRef(id)
}
\arguments{
\item{id}{EA/SEPA/COSMOS station id.}
}
\value{
ref string to request flow from API
}
\description{
For EA/SEPA/COSMOS stations for which data is provided, this provides the
reference string for the API, given the
 identifier.
}
\examples{
\dontrun{
idToRef("SS50F007")
}

}
