% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfInterface-package.R
\docType{package}
\name{rfInterface}
\alias{rfInterface}
\alias{rfInterface-package}
\title{Functions for Using UK Riverflow and Rainfall API}
\description{
rfInterface: Functions for Using UK Riverflow and Rainfall API
}
\details{
rfInterface provides functions to easily access EA and NRFA river
flow data and SEPA rainfall data, with utility functions to convert station
identifiers and location grid references. Preview of data at
https://gateway-staging.ceh.ac.uk/hydrology-ukscape.
}
